move to kesselinventoryservice.go
